refactor(app): move OmiConfirmDialog to widgets, remove lib/ui folder
Relocate the last remaining file in lib/ui (molecules/omi_confirm_dialog.dart)
to lib/widgets alongside the app's other dialogs, update the five importers,
and delete the now-empty lib/ui folder left over from the discarded
design-system adoption.
