+++ app/lib/pages/conversations/auto_sync_page.dart
@@ -9,7 +9,7 @@ import 'package:omi/pages/conversations/private_cloud_sync_page.dart';
 import 'package:omi/providers/sync_provider.dart';
 import 'package:omi/providers/user_provider.dart';
 import 'package:omi/services/wals.dart';
-import 'package:omi/ui/molecules/omi_confirm_dialog.dart';
+import 'package:omi/widgets/omi_confirm_dialog.dart';
 import 'package:omi/utils/l10n_extensions.dart';
 import 'package:omi/utils/other/temp.dart';
 import 'package:omi/utils/sync_confirmation.dart';
